Refactor: move client logic outside page.tsx
