+++ src/app/[locale]/incident/contact/components/IncidentContactPage.tsx
@@ -1,0 +1,21 @@
+'use client'
+
+import { useTranslations } from 'next-intl'
+import { useFormStore } from '@/store/form_store'
+import { Heading } from '@/components'
+import { IncidentContactForm } from '@/app/[locale]/incident/contact/components/IncidentContactForm'
+
+export const IncidentContactPage = () => {
+  const t = useTranslations('describe-contact')
+  const { loaded } = useFormStore()
+
+  // TODO: implement nice loading state, for if loaded if false. Also create a loading state for Stepper store.
+  if (loaded) {
+    return (
+      <div className="flex flex-col gap-4">
+        <Heading level={1}>{t('heading')}</Heading>
+        <IncidentContactForm />
+      </div>
+    )
+  }
+}
