+++ src/app/[locale]/incident/contact/page.tsx
@@ -1,21 +1,5 @@
-'use client'
-
-import { useTranslations } from 'next-intl'
-import { IncidentContactForm } from '@/app/[locale]/incident/contact/components/IncidentContactForm'
-import { Heading } from '@/components/index'
-import { useFormStore } from '@/store/form_store'
+import { IncidentContactPage } from '@/app/[locale]/incident/contact/components/IncidentContactPage'
 
 export default function AddContactDetailsPage() {
-  const t = useTranslations('describe-contact')
-  const { loaded } = useFormStore()
-
-  // TODO: implement nice loading state, for if loaded if false. Also create a loading state for Stepper store.
-  if (loaded) {
-    return (
-      <div className="flex flex-col gap-4">
-        <Heading level={1}>{t('heading')}</Heading>
-        <IncidentContactForm />
-      </div>
-    )
-  }
+  return <IncidentContactPage />
 }
