+++ src/app/[locale]/incident/summary/page.tsx
@@ -1,21 +1,5 @@
-'use client'
-
-import { useTranslations } from 'next-intl'
-import { IncidentSummaryForm } from '@/app/[locale]/incident/summary/components/IncidentSummaryForm'
-import { Heading } from '@/components/index'
-import { useFormStore } from '@/store/form_store'
+import { IncidentSummaryPage } from '@/app/[locale]/incident/summary/components/IncidentSummaryPage'
 
 export default function SummaryDetailsPage() {
-  const t = useTranslations('describe-summary')
-  const { loaded } = useFormStore()
-
-  // TODO: implement nice loading state, for if loaded if false. Also create a loading state for Stepper store.
-  if (loaded) {
-    return (
-      <div className="flex flex-col gap-4">
-        <Heading level={1}>{t('heading')}</Heading>
-        <IncidentSummaryForm />
-      </div>
-    )
-  }
+  return <IncidentSummaryPage />
 }
